# ViagemAPI/ViagemAPI/ViewModel/MotoristaViewModel.cs
namespace ViagemAPI.ViewModel
{
    public class MotoristaViewModel
    {
        public int Id { get; set; }
        public string Nome { get; set; }
        public string Cpf { get; set; }
    }
}